pyproject.toml
[project]
name = "my-api-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.15.3",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]
